Fixed EllPath init when it is not set
First, it includes “.”, then ~/lib/ell exists, it is added, and
finally, if GOPATH is set, then the checked out directory is added.
This works with a default “go get” of gell, if you leave the GOPATH the
same.

You can always set the ELL_PATH environment variable to force it to
what you want. The ell.ell file must be found in the path for gell to
boot.