added keys: option to read, more cleanup.
sets the type of keys for structs. An attempt is made to coerce keys to
that. The default is Null, so normal read (and json decode) will return
whatever is there. Setting to <keyword> forces keyword keys, <string>
forces string keys, and <symbol> forces symbol keys. Both symbols and
keywords have constraints,  may not succeed.